feat(...) : ChattingFragment 에 기본 socket 세팅
