+++ app/src/main/java/com/iron/espresso/presentation/home/mystudy/studydetail/chat/ChatAdapter.kt
@@ -1,0 +1,28 @@
+package com.iron.espresso.presentation.home.mystudy.studydetail.chat
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
+import com.iron.espresso.R
+
+class ChatAdapter : ListAdapter<ChatItem, ChatViewHolder>(DIFF_CALLBACK) {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ChatViewHolder {
+        return ChatViewHolder(parent)
+    }
+
+    override fun onBindViewHolder(holder: ChatViewHolder, position: Int) {
+        holder.bind(currentList[position])
+    }
+
+    companion object {
+        private val DIFF_CALLBACK = object : DiffUtil.ItemCallback<ChatItem>() {
+            override fun areItemsTheSame(oldItem: ChatItem, newItem: ChatItem): Boolean =
+                oldItem.message == newItem.message
+
+            override fun areContentsTheSame(oldItem: ChatItem, newItem: ChatItem): Boolean =
+                oldItem == newItem
+        }
+    }
+}
